# openapi.yaml
openapi: 3.1.0
info:
  title: GMAO API
  description: API Proxy - Gateway entre frontend et données
  version: 2.2.1
paths:
...
  /users/:
    get:
      tags:
        - users
      summary: List Users
      description: Liste les utilisateurs avec filtres optionnels
      operationId: list_users_users__get
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            description: Offset de pagination
            default: 0
            title: Skip
          description: Offset de pagination
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 1000
            minimum: 1
            description: Nombre max de résultats
            default: 100
            title: Limit
          description: Nombre max de résultats
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            description: Filtrer par statut (active, suspended, etc.)
            title: Status
          description: Filtrer par statut (active, suspended, etc.)
        - name: search
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: "null"
            description: Recherche sur nom, prénom, email
            title: Search
          description: Recherche sur nom, prénom, email
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserListItem"
                title: Response List Users Users  Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /users/{user_id}:
    get:
      tags:
        - users
      summary: Get User
      description: Détail d'un utilisateur par ID
      operationId: get_user_users__user_id__get
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            title: User Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserListItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        first_name:
          anyOf:
            - type: string
            - type: "null"
          title: First Name
        last_name:
          anyOf:
            - type: string
            - type: "null"
          title: Last Name
        email:
          anyOf:
            - type: string
            - type: "null"
          title: Email
        initial:
          anyOf:
            - type: string
            - type: "null"
          title: Initial
        status:
          type: string
          title: Status
          default: active
        role:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Role
      type: object
      required:
        - id
      title: UserListItem
      description: Schema léger pour listes d'utilisateurs
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        first_name:
          anyOf:
            - type: string
            - type: "null"
          title: First Name
        last_name:
          anyOf:
            - type: string
            - type: "null"
          title: Last Name
        email:
          anyOf:
            - type: string
            - type: "null"
          title: Email
        location:
          anyOf:
            - type: string
            - type: "null"
          title: Location
        title:
          anyOf:
            - type: string
            - type: "null"
          title: Title
        description:
          anyOf:
            - type: string
            - type: "null"
          title: Description
        tags:
          anyOf:
            - {}
            - type: "null"
          title: Tags
        avatar:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Avatar
        status:
          type: string
          title: Status
          default: active
        role:
          anyOf:
            - type: string
              format: uuid
            - type: "null"
          title: Role
        initial:
          anyOf:
            - type: string
            - type: "null"
          title: Initial
        last_access:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
          title: Last Access
      type: object
      required:
        - id
      title: UserOut
      description: Schema complet pour détail utilisateur
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError